Implement auto-hashing for cache-to-disk moves
When the Unraid Mover runs, a partial file is created on the disk. This file.partial is then moved to the original file name on the disk. Therefore, inotifywait does not see the move as a close_write, but rather a moved_to.

Note that this hotfix will cause normal moves on the disk (i.e. those using the "mv" command) to trigger a rehash. I will address that in the future.